expense-report/requirements.txt
xlrd>=2.0.2
openpyxl>=3.1.0
Pillow>=10.0.0
pytest>=7.0.0
